Replace Fred with redis-rs

+++ nativelink-config/src/stores.rs
@@ -1150,6 +1150,7 @@ pub struct RedisSpec {
     ///
     /// Default: 4096
     #[serde(default, deserialize_with = "convert_numeric_with_shellexpand")]
+    // DEPRECATED
     pub broadcast_channel_capacity: usize,
 
     /// The amount of time in milliseconds until the redis store considers the
@@ -1203,7 +1204,7 @@ pub struct RedisSpec {
     ///
     /// Default: 10000
     #[serde(default, deserialize_with = "convert_numeric_with_shellexpand")]
-    pub scan_count: u32,
+    pub scan_count: usize,
 
     /// Retry configuration to use when a network request fails.
     /// See the `Retry` struct for more information.
